forgot and registration controller move in security controller
